requirements.txt
tqdm
gradio
pandas
openpyxl
Pillow
matplotlib
numpy
six
pycocotools
shapely
opencv-python
